+++ src/paritea/glue/stim/__init__.py
@@ -1,4 +1,4 @@
-from .detector_error_model import export_to_stim_dem, push_out_for_measurement_detectors
+from .dem import export_to_stim_dem, push_out_for_measurement_detectors
 from .sinter import error_rates_equal, wrap_dem_as_sinter_task
 from .stim_circuit import from_stim
