+++ src/paritea/glue/stim/dem.py
